split up Register.js, text fields are child components with different states
